requirements.txt
easyocr
Pillow
discord.py
python-dotenv